admin request users page is added

+++ blueprints/user.py
@@ -169,20 +169,22 @@ def append_service(id):
     flash("ثبت درخواست با موفقیت انجام شد", "success")
     return redirect(url_for("user.dashboard"))
 
-@bp.route("/user/dashboard/delete-service/<int:id>")
+@bp.route("/user/dashboard/delete-service/<int:id>", methods=["GET", "POST"])
 @login_required
 def delete_service(id):
-    item = ServiceRequest.query.filter(
-        ServiceRequest.id == id,
-        ServiceRequest.user_id == current_user.id,
-        ServiceRequest.status == "pending"
-    ).first_or_404()
     
-    db.session.delete(item)
-    db.session.commit()
+    if request.method == "POST" :
+        item = ServiceRequest.query.filter(
+            ServiceRequest.id == id,
+            ServiceRequest.user_id == current_user.id,
+            ServiceRequest.status == "pending"
+        ).first_or_404()
+        
+        db.session.delete(item)
+        db.session.commit()
 
-    flash("سرویس با موفقیت حذف شد", "success")
-    return redirect(url_for("user.dashboard"))
+        flash("سرویس با موفقیت حذف شد", "success")
+        return redirect(url_for("user.dashboard"))
 
 
 @bp.route("/user/dashboard/edit-service/<int:id>")
